requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangGraph & LangChain (compatible versions)
langgraph>=0.0.43
langchain>=0.1.4
langchain-google-genai>=0.0.6
langchain-core>=0.1.46

# Google Generative AI
google-generativeai>=0.3.2

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Environment Management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Zep Context AI Memory
zep-cloud>=2.0.0
